fix: 音量设置最小时文本显示异常
音量设置最小时，用以显示文本的数值异常

Log: 修改音量设置最小时文案异常问题
Bug: https://pms.uniontech.com/zentao/bug-view-47091.html
Reviewed-on: http://gerrit.uniontech.com/c/dde-control-center/+/4199
Tested-by: <mailman@uniontech.com>

+++ src/frame/window/modules/sound/speakerpage.cpp
@@ -273,7 +273,7 @@ void SpeakerPage::initSlider()
         m_speakSlider->blockSignals(true);
         m_speakSlider->setValue(v * 100 + 0.000001);
         m_speakSlider->blockSignals(false);
-        m_outputSlider->setValueLiteral(QString::number(v * 100 + 0.000001) + "%");
+        m_outputSlider->setValueLiteral(QString::number((int)(v * 100 + 0.000001)) + "%");
     });
 
     connect(m_model, &SoundModel::maxUIVolumeChanged, this, [ = ](double maxvalue) {
@@ -292,7 +292,7 @@ void SpeakerPage::initSlider()
         m_speakSlider->blockSignals(true);
         m_speakSlider->setValue(static_cast<int>(m_model->speakerVolume() * 100 + 0.000001));
         m_speakSlider->blockSignals(false);
-        m_outputSlider->setValueLiteral(QString::number(m_model->speakerVolume() * 100 + 0.000001) + "%");
+        m_outputSlider->setValueLiteral(QString::number((int)(m_model->speakerVolume() * 100 + 0.000001))+ "%");
     });
 
     m_layout->insertWidget(3, m_outputSlider);
